Build gtest using stlport.
src/Android.mk
Added _host prefix to the module name for host build to prevent
name clashes with the device's targets.

tests/Android.mk
Added _host prefix to the tests built on host.
Use new variables to list the required libraries. This is needed
because the host libs now have a _host suffix so I maintain the 2
lists of libs separatly.
Tests are systematically built.

Tested: On linux, build passion-eng and sim-eng

Needs to be submitted before Id4771c5f