Build distribution script for splash2d.
Added a script which builds a distribution source tree for splash2d.
This performs the following:
* Removes files that aren't required / tested.
* Builds redistributable Visual Studio and Xcode projects on Windows and Mac
  machines.
* Packages the results in zip and tar.gz archives.

There are a few caveats of using cmake to generate Visual Studio and Xcode
solutions.
* Pre / post build scripts and install steps can't be used as they pull in
  cmake dependencies.
* The process to used to convert absolute to relative paths is fairly brittle
  so it wouldn't be surprising if small changes in Visual Studio or Xcode
  file formats broke this.
* cmake authors do not endorse this use case
  http://www.cmake.org/Wiki/CMake_FAQ#Why_does_CMake_use_full_paths.2C_or_can_I_copy_my_build_tree.3F

Tested:
Built distribution package, copied to Linux, Windows and OSX hosts, unpacked
and verified that it's possible to build and run the debug targets.

Bug: 11043660

Change-Id: I6da22649d24049ab4e4830e0fcbe95134e6214a5