Created Android NDK projects for each unit test executable.
Unit tests have been moved into separate subdirectories so that it was
possible to create an Android NDK project for each test that can be
built using ndk-build and "android update project".

Android project files for each unit test are generated from the template
files under AndroidTemplate using the AndroidTemplate/generate_test_project.sh
script.

At the moment the HelloWorldTests application is not tested on Android as it
requires additional functionality to read a golden file from the apk.

Also, AndroidUtil/build_apk.sh has been modified so that it's possible
to deploy and launch an application independently of the build step in addition
to fixing an erroneous error message when only one Android device is connected.

Tested:
On Android...
$ ./Box2D/Unittests/run_tests_android.sh
... lots of output ...
The following tests passed (4/4):
  ConfinementTests
  ConservationTests
  CallbackTests
  FunctionTests

On Linux...
$ cd Box2D
$ cmake -G'Unix Makefiles' && make -j16
$ ./Unittests/run_tests.sh
.. all test pass ..

On Windows...
% cd Box2D
% cmake -G"Visual Studio 11"
% msbuild.exe Box2D.sln
% Unittests\run_tests.bat
.. all test pass ..

On OSX...
$ cd Box2D
$ cmake -GXcode && xcodebuild
$ ./Unittests/run_tests.sh
.. all test pass ..

Change-Id: Iad228c9c3c2b9440be3d62af680e162cfbf5d1f6