Added a script to build / deploy / execute a NDK application.
build_apk.sh simplifies building, deploying and executing NDK applications.
build_apk.sh is based upon
Box2D/{HelloWorld,Testbed}/android_build_debug_apk.sh.  It generalizes the
scripts to make it possible to build any NDK project and adds the default
behavior of running the application and streaming the log output to the
terminal.

Tested:
For example, to build the release HelloWorld package, deploy it to the
device with serial number 06d8bd43, execute the application and displays the
application's log output to the terminal...

$ cd HelloWorld
$ build_apk.sh ADB_DEVICE=06d8bd43

To perform a debug build and start an application with GDB...

$ cd HelloWorld
$ build_apk.sh ADB_DEVICE=06d8bd43 NDK_DEBUG=1 RUN_DEBUGGER=1

To build a release apk without deploying or launching the package...

$ cd HelloWorld
$ build_apk.sh DEPLOY=0 LAUNCH=0

Bug: 10550662
Change-Id: Ib26f9eee6a051652834f71245b0858846791dff3