Made it possible to build googletest from ndk-build.
In addition, redirected test output to the Android log stream so that it's
visible using "adb logcat" and added the ability to redirect test output
to custom printf functions on Android.

Tested:
In conjunction with I78959be33d8c8d7e906f3b577094cf37ad81a96d verified all
unit tests (with the exception of HelloWorldTests which is a work in progress)
send output to the Android log stream.

Bug: 10550662
Change-Id: I8c83ca2ff62ee057e7d157d460461f08bd047f97