Fixed build_apk.sh target SDK selection and install + deploy on 2.3 devices.
The script now detects the target SDK based upon the SDKs installed on the
host with the constraint that the selected SDK is newer than the version
specified by the BUILDAPK_ANDROID_TARGET_MINVERSION variable.

Since "adb install" can fail and not return an error code, worked around the
issue by scraping the command's output.

"am start" doesn't support the -S argument (stop running activity) on 2.3
devices so this behavior is disabled on Android 2.3.x.

In addition...
* Wrapped ndk-build with a function that determines whether ndk-build is in the
  PATH and if it isn't tells the user to add the NDK dir to the PATH.
* Wrapped android with a function that determines whether android is in the
  PATH and if it isn't tells the user to add the sdk/tools dir to the PATH.
* Skipped package build if AndroidManifest.xml doesn't specify a package
  name.  This is useful when building the splash2d library from the
  splash2d/Box2D directory.

Tested:
Verified that it's possible to build, deploy and launch Testbed on a
Samsung Galaxy Y running Android 2.3.
$ cd Box2D/Testbed
$ build_apk.sh APP_ABI=all

Bug: 11436254,11436188,11392134,11420177
Change-Id: Ie0d3c7fe038e3b58e079c95e459265c9b6325e7c