Filtered unused cmake generated projects from Visual Studio solution.
cmake generates ALL_BUILD, ZERO_CHECK and INSTALL projects which are not
required in the distribution build of splash2d.  This removes these projects
from the tree and references to them from the solution file.  Since there
are now no dangling references to projects this makes it possible to build
using msbuild.exe.

Also, removed .git* files from the distribution tree and the Autobuild
directory.

Tested:
Built a splash2d distribution for Windows and verified that it's possible to
build using the Visual Studio GUI and msbuild.exe.

Bug: 11337156
Change-Id: I44fabd13c683f5a232a300199ea62ffe2bd3ce4f