Updated Building.txt to specific a minimum cmake version for OSX 10.9.
Generating the Xcode project on OSX 10.9 (Mavericks) with cmake version
2.8.11.x or below results in the following error...

CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
OPENGL_INCLUDE_DIR (ADVANCED)
   used as include directory in directory /Users/smiles/splash2d_tree/vendor/unbundled_google/libs/freeglut
   used as include directory in directory /Users/smiles/splash2d_tree/vendor/unbundled_google/libs/freeglut
   used as include directory in directory /Users/smiles/splash2d_tree/vendor/unbundled_google/libs/splash2d/Box2D/glui
   used as include directory in directory /Users/smiles/splash2d_tree/vendor/unbundled_google/libs/splash2d/Box2D/Testbed

-- Configuring incomplete, errors occurred!

The resolution for this is included in cmake 2.8.12.1.

Tested:
Verified that it's possible to generate the Xcode project on OSX 10.9.
Successfully built the Xcode project, ran all unit tests and sample
applications.

Change-Id: I72e36d3db18026c10fada56e4811a1bd3ad3c7a6