Script which merges multiple git projects into a single repo.
push_upstream.sh takes a set of git projects and places each project in a
subdirectory of a unified git project while retaining the version history
of all projects.

For example, if projects A B and C are provided they're placed in a git project
under subdirectories ${root}/A ${root}/B ${root}/C and merged into a single
project.  This is achieved by cloning each project and rewriting each commit
to move the files into a subdirectory.

In addition, push_upstream.sh provides an option to scrub e-mail addresses
and names from commits replacing them with anonymous@google.com and
Anonymous respectively.

Tested:
Verified that it's possible to merge the liquidfun, googletest and freeglut
projects into a single project.  Note splash2d is in the process of being
renamed to liquidfun.

$ ./tools/push_upstream.sh -e '(XXX@google.com)' -a '(YYY)' -r \
  "$(for r in splash2d freeglut googletest; do \
     echo "https://${internal_path}/${r}|${r}|${internal_branch}"; done)"
...
$ cd push_upstream*
$ ls
freeglut googletest splash2d
$ git log
$ git --grep (XXX|YYY)
... verified history does not contain XXX or YYY ...
$ git log
... manually inspected the history ...

Bug: 11043179
Change-Id: I6c68b79427d8d13054a2601056bfa168e76f4c40