Workaround the introduction of the liquidfun mirror project.
This is a temporary work around the liquidfun project being present in the
tree which results in multiple libliquidfun and libliquidfun_static build
targets being visible to the Android build system.

Tested:
Verified that it's possible to build everything under the splash2d and
liquidfun projects using the Android build system.

$ bash
$ source build/envsetup.sh
$ tapas nada
$ make libsplash2d libsplash2d_static libliquidfun libliquidfun_static
.. success ..

Bug: 11880022
Change-Id: I5b66e08ba4bd4dfc07c86ba68e495d1e16b408e4