Made particles a lot prettier on Android using blended gaussian points
Tested: on a wide range of Android devices

(for "dambreak", on old Nexus 7, slowest device):

pretty particles: 35 msec
square particles - correct size: 35msec
square particles - 2 pixels (current): 33 msec

seems we are mostly cpu bound on the old nexus 7, so this effect is for free.

on Nexus 4, new Nexus 7, nexus 10, etc this is all 60fps or over.

* refactored particle rendering code
* added FPS display
* added minimal version of this on desktop platforms

Change-Id: I6ae6b5b9201075394bdd6497b7bdd83c31447def