Compute depth on next frame.
(1) When a group is created or joined,
the particle system will no longer compute depth immediately,
but will set a flag to do so on next step.
(2) Set the same flag when the member particles are deleted
or when the flag b2_solidParticleGroup is switched by SetGroupFlag().
(3) Store internal and temporary flags (e.g. m_toBeDestroyed) in m_groupFlag.

Bug: 11481884
Change-Id: Ib0ebe38cad48d800c5532cb9f14a82aeb7e4e46e
Tested: Testbed and Unittests on OSX