Added a particle flag to prevent other particles from leaking.
This patch adds b2_barrierParticle as a b2ParticleFlag.
This flag must be used together with b2_wallParticle as follows:
    b2_wallParticle | b2_barrierParticle;
If a particle is colliding with a line segment connecting paired barrier particles,
its velocity will be decelerated to avoid the collision.

This patch also enables to create a particle group from b2EdgeShape or
b2ChainShape for unit tests of barrier particles.

Tested: DrawingParticle in Testbed and ConfinementTests in Unittests on OSX
Bug: 11421729
Change-Id: I4c98c05696c6431741f0a031584dea526358170c