Fixed HelloWorldTests so that it finds the golden file on OSX and Windows.
Fixed a bug in BodyTracker::SetWorkingDirectory() to the directory containing
the executable on Windows and OSX.

Windows and OSX builds result in the executables being placed under
UnitTests/{Debug,Release} which differs to the Linux build.  So this
change just takes the executable path and sets the baselines directory to
the container directory.

Tested:
Verified HelloWorldTests passes when executed...
* On Windows from Visual Studio.
* On Windows from Unittests/run_tests.bat
* On Linux from Unittests/run_tests.sh
* On OSX from Unittests/run_tests.sh
* On OSX from Xcode.

Bug: 12015991
Change-Id: I8f0b538be590f8f9e8c8869b56d0db401c0fb8d8