Reorder operations in b2ParticleSystem::Solve.
This patch reorders operations in b2ParticleSystem::Solve so that
particle positions will be updated at the tail of the function
and then particle velocities will express precisely the last movements / dt.

This patch affects collisions between particles and bodies.
Velocity of a body which is colliding with particles is now limited
for numerical stability.

Bug: 11552012
Tested:Unittests and Testbed on OSX

Change-Id: I864411a71ec78c2396ec2e5c80572c0739bb7639