Adding new test, Sandbox.
Contains new test, a maze of flowing sand filled with pumps.  It's designed
so that people playing with the app can easily reorganize the maze of pumps
and blocks.  It demonstrates killfields, prismatic joints, and collisions with particles
driven by forces besides gravity.

Tested:
Verified that Sandbox test executes correctly in the
Testbed application on Android.

Change-Id: Ie2782e5ff479fa3cbd8eee124dab0bfed932de78