Rescale tests featuring particles in Testbed to 1/10.
The purpose of this patch is to make object sizes more similar to the sizes of
natural objects. Particle sizes will be 2.5-3.5cm, and world size will be 4m.

Smaller objects may cause explosion, vibration or other strange behavior of
particles. So this patch contains some improvements in the numerical stability.
* Set particleIterations as 3 by default in Testbed.
* Tensile and powder force no longer act between a particle and a body.
  (There are still pressure and damping force.)
* Reduce reaction force to bodies with small density.
* Improved the way to update static pressure.
* Added extra damping force between bodies and b2_staticPressureParticle.

Tested: Testbed on OSX and Android
Bug: 11325217

Change-Id: I494e458ca8ff92b5b1022a59f9b95dc6226b2fb0