Adding code coverage to liquidfun unit test.
Added a check to see if lcov is installed if not,
displayed a text to ask the user to install lcov.
Removed redandacy from run_tests.sh.
Removed some white spaces from CMakeLists.txt.
*sigh* removed extra blank lines in CMakeLists.txt.
Made changes suggested in the review by smiles@.
Added the full path for the code coverage index.html.
Added a check for Clang compiler in CMakeList.txt
Minor indentation fixes.
Tested:
 - Run it on linux and verified the code coverage generated files are in the codeCoverage directory
 - Run it on Mac and verified the code coverage generated files are in the codeCoverage directory

P.S left the processing noise for now, I'll add an option for that in a bit!

Change-Id: I148dd1c535b53c134296a07fae9c795d0d23e4cd