Improved ./tools/add_copyright.sh documentation.
* At the top of the script described how to access the help text.
* Described which files are modified by the script.
* How files are modified by the script.
* What the script arguments do.
* Provided an example of how to use the script.

Tested:
Ran './tools/add_copyright.sh -h' on Linux + OSX and verified the output.

On my machine it prints out....

Add a copyright header to *all* Google authored and modified source files.

This script adds a copyright header to all files found in...
  /XXXX/AAA/BBB/CCC/DDD/liquidfun
... that have been modified in this project's git repository with respect to
the specified upstream git repository.

This script uses git to perform the diff between the upstream repository and
this project so it requires this project to be a valid git workspace.

In the case of files with no copyright header the following header is added:
"
Copyright (c) 2014 Google, Inc.

This software is provided 'as-is', without any express or implied
warranty.  In no event will the authors be held liable for any damages
arising from the use of this software.
Permission is granted to anyone to use this software for any purpose,
including commercial applications, and to alter it and redistribute it
freely, subject to the following restrictions:
1. The origin of this software must not be misrepresented; you must not
claim that you wrote the original software. If you use this software
in a product, an acknowledgment in the product documentation would be
appreciated but is not required.
2. Altered source versions must be plainly marked as such, and must not be
misrepresented as being the original software.
3. This notice may not be removed or altered from any source distribution.

"

Where files already have a copyright header and have been modified by Google
the following is inserted in the header:
"
Copyright (c) 2014 Google, Inc.
"

Source files matching the following regular expressions are modified:

  "\.cpp$"
  "\.c$"
  "\.h$"
  "\.mk$"
  "\.sh$"
  "\.bat$"
  "CMakeLists\.txt$"
  "/\.project$"
  "\.xml$"

Usage: add_copyright.sh [-g upstream_git_repo] [-b upstream_git_branch]

-g upstream_git_repo:
  upstream_git_repo is the location of the upstream git repository to compare
  against.  If this isn't specified it defaults to
  /XXXX/AAA/BBB/CCC/DDD/../../../../external/box2d

-b upstream_git_branch:
  upstream_git_branch specifies the branch in the upstream_git_repo to compare
  against.  If this isn't specified it defaults to master.

For example, assuming we're using the default git upstream repo and branch:
$ cd /XXXX/AAA/BBB/CCC/DDD/liquidfun
$ pushd ../../../../external/box2d
$ git checkout -b master origin/master
$ popd
$ ./tools/add_copyright.sh
$ git status -s
\# Displays a list of modified files.

Bug: 12309330
Change-Id: I8398e31ee9358afd6e38cb4bdec620983010c11e