Fixed wrap behavior of particle parameter increment / decrement.
Particle parameters were not correctly wrapping around 0 as
(uint32)MAX_UINT % N != (int32)MAX_INT % N.  So this changes the behavior
of ParticleParameter to keep indices in the range 0..N where N is the number
of parameters in the parameter definition.

Tested:
Verified that decrementing a parameter wraps to N when the initial value is 0
in the Testbed on Linux.

Change-Id: I3773d568fcd44c351e3e21c8993131f3cce60484