Fixed uninitialized variable in Testbed ParticleParameter.
ParticleParameter::Changed() would potentially never set "restart" which is
uninitialized in SimulationLoop() in Main.cpp leading to undefined behavior.
Specifially on Nexus 7 (1st gen) it just so happened that a non-zero value up
ended on the stack at this point so the test would be stuck continually
restarting.

Tested:
Verified that Testbed doesn't continuously restart tests on Nexus 7 (1st gen).

Bug: 12803979
Change-Id: I502fce28f0c2596823ce8642ff831d39752c05e0