Allow multiple b2ParticleSystems in b2World
b2ParticleSystems are created and destroyed with new APIs:
- b2World::CreateParticleSystem
- b2World::DestroyParticleSystem
These APIs match the existing b2World pattern: e.g. CreateBody, DestroyBody.

Note that we haven't yet deleted the old API in b2World--the one that assumes
there is only one b2ParticleSystem. We will delete that API in a future change.
For now, the old API just uses the first b2ParticleSystem. Also, we create a
default particle system when b2World is created.

Bug: 11278941
Tested: Change is a nop so not a lot to test. Built and ran Testbed and
Unittests on Linux for Android.
Change-Id: I5031add9a0908d0f0bd56a5b78af61e9ec154307