Fixed a bug due to empty particle group.
m_depthBuffer is allocated the first time b2_solidParticleGroup is used.
However, if the particle capacity is zero at that time, the block allocator
will return NULL and then the buffer will be treated as not allocated.

This bug is actualized when you create a group with b2_solidParticleGroup and
no particles.

This patch fixes this bug by setting non-zero value as the particle capacity
before allocating m_depthBuffer.

This patch also modifies DestroyParticleGroup() to use SetParticleGroupFlag()
to properly process flags of destroyed groups.

Tested: CreateParticleGroupWithNoShape in FunctionTests
Bug: 12925151
Change-Id: Ibf1555c11b964094a16c95b8ac7f9ee79b063e74