Optionally detect "stuck" particles
Allow a user to set a threshold (in iterations) after which a
particle is considered a candidate to be "stuck" if it is continuously
contacting multiple fixtures per step.

Adds an API to set the threshold, and query the per-step "stuck"
list.

Also included is a minor refactor of the growable arrays to
eliminate copypasta.

Unit test added to test stuck particle detecting with various
thresholds. BodyContactTest refactored to reflect multiple particle
system support as well, and test cases updated.

Tested on OSX, Linux, Android.

Bug: 11552217
Change-Id: Id160001ec137e7fb2e5f5240558c264db7b26483