Call all particle-systems in b2World::Step, QueryAABB, and RayCast
This fixes a bug in the multiple-particle-system change. I was only
checking the first particle system in these functions.

Also, since multiple particle systems can exist, the callback needs
to pass in the particle system that corresponds to the particle
index, in the callback.

Tested: Testbed and Unittests on Android.
Bug: 11278941 Add support for multiple particle systems
Change-Id: I0941fec2f49ef2da07468ed26cbbad88aa9c1a55