Add option to cull an entire ParticleSystem from RayCast.
b2ParticleSystem::RayCast can be expensive, especially when no
particles are hit. Prior to this change, we'd only early-out of
b2ParticleSystem::RayCast when (1) a particle was hit, and
(2) the ReportParticle callback returned 0. So, even if the
b2RayCastCallback has no interest in casting against particles,
many cycles may still be consumed.

The fix is to add an API that asks the callback if it's
interested in a particle system before ray casting against it.

Change-Id: Ic8d542e7b25033ea9dae74c6cd486eb02e470630
Tested: Testbed and unit tests. Updated RayCast unit test to check
new b2RayCastCallback::ShouldQueryParticleSystem API.
Bug: 11159187 Option to disable raytracing against particles