Allow entire particle systems to be culled from b2World::QueryAABB.
We use a callback API (b2QueryCallback::ShouldQueryParticleSystem)
so that the caller has control over which particle systems are
included in the query.

By default, all particle systems are included. That is, the default
behavior is unchanged.

This new API can be a significant optimization. Previously, the only
way to early-out of b2ParticleSystem::QueryAABB was to return false
from b2QueryCallback::ReportParticle. Hence, if no particles were
ever reported, there was no way to early out.

Tested: Added unit test to Callbacks. Ran unit tests and testbed on
Android.
Bug: 13108767 Allow particle systems to be removed from QueryAABB,
RayCast, and Step calls

Change-Id: I08013abf2c1d57dd53a20d3d8d001c6b4fb7518e