Add ApplyLinearImpulse and expose ApplyForce in b2ParticleSystem.
Allows an impulse or force to be applied to an individual
particle, or to a group of particles.

Also add the same functions to b2ParticleGroup, to apply impulse or
force to the entire group.

These functions are similar to b2Body::ApplyLinearImpulse and
b2Body::ApplyForce, except that we apply the impulse and force to
the center of each particle.

Bug: 11552220 Make it possible to add impulses to particle groups
Change-Id: I6a09ad73627a159adc8dfbad3a260d1f8c968dad
Tested: Added "Impulse" to testbed. Ran testbed and unit tests on
Android.