Modify build_apk.sh to monitor the log until an app launches or times out.
This workaround is required since "am start -W" can end up waiting
indefinitely for an application to start in cases when an app is launched,
moved to the background by the activity manager and finally destroyed.

Tested:
Ran the following (on Linux and OSX) and verified build_apk.sh never hangs
when launching unit tests on Nexus 7 (gen2) and Nexus 10.

$ cd Box2D/Unittests
$ for i in {0..100}; do ./run_tests_android.sh; done

Bug: 12020116
Change-Id: Ia9fa149ae66fb2d2a66f41762501dc4599eced95