Added Soup Stirrer test to the Testbed.
This test adds a dynamic body to "Soup" and continuously applies a force to
the body to agitate the fluid.  By default the motion of the body is
restricted the x-axis using a prismatic joint.  The joint restricting the
body can be enabled / disabled using "t" or by clicking / tapping in the soup.

In addition, updated the release notes with a mention of this new test.

Tested:
Verified the test works on Linux, OSX, Windows and Android.

Bug: 12176842
Change-Id: I736a3acc73581e439f380f46ab5c74e701d12604