Added Maxwell test / game to the Testbed.
This is a game which adds some fun to Maxwell's demon.
http://en.wikipedia.org/wiki/Maxwell's_demon
The user's goal is to try to catch as many particles as possible in the
bottom half of the container by splitting the container using a barrier
with the 'a' key or by tapping inside the container.

Tested:
Verified it's "playable" on Windows, OSX, Linux and Android.

Bug: 12176842
Change-Id: I47195f546d71f01e83bdcfec53850ec5c6af3bb3