Remove 'Particle' from many of b2ParticleSystem's function names.
Moved b2ParticleSystem::DeleteParticlesInGroup to
b2ParticleGroup::DeleteParticles.

Renamed PauseSimulation and IsSimulationPaused to SetPaused and
GetPaused.

Renamed the private class ParticleBuffer to UserOverridableBuffer,
and renamed the private function RequestParticleBuffer to
RequestBuffer.

All of the old function names work for the moment. I've included
re-direct functions so that we can run with both the old and the
new names at the same time. Once all the clients have been
upgraded, I'll delete the re-direct functions.

Tested: Unittests and testbed on Android and Linux.
Bug: 13161203 Methods in b2ParticleSystem should drop the "Particle"
prefix in their names

Change-Id: I924dd575efefd1b07da3415ae053d6397d845db4