Allow small calculation error in the strict contact check.
The strict contact check uses TestPoint() to determine whether to remove a
contact. However, the function is very sensitive to numerical error especially
when it is called for points near the boundary. In addition, it always returns
false if the shape is a edge or chain shape.

This patch modifies the strict contact check to allow small error. That reduces
instability of interactions between particles and bodies and enables edge and
chain shapes to expectedly interact with particles.

Bug: 13347881
Bug: 13304312

Change-Id: I988e1bba207b6d95a182e9e1deb3b7607820c730