Reduce number of particles in some tests
The previous change to reduce the number of particles in Testbed
tests missed the tests AntiPointy, Pointy and Ramp due to concurrent
committing.  This increases particle size and reduces particle number
in those tests.

This has the added benefit of making the bug leading to the strict
particle contact check even mroe apparent in Ramp.

Change-Id: I478226c9c5fc867dd1e7f5574a63334a6f772117
Tested: ran Testbed on Linux and Android
Bug: 13419137