10s timeout for log monitoring when launching apps with build_apk.sh
The activity manager on Gingerbread devices doesn't support -n (wait for
activity to start) and there don't appear to be deterministic messages from
logcat to indicate that an app launch is complete.  Since there appears
to be no signal to determine whether an app launch is complete this modifies
build_apk.sh to timeout log monitoring after 10 seconds when deploying to
devices that run Gingerbread.

Tested:
Verified that all Box2D/Unittests/run_tests_android.sh works with the Samsung
Galaxy Y and the Nexus 10.

Bug: 13413185
Change-Id: I080e07ad24e49c616cdbddc149392098deabc1d8