Add b2ParticleSystem to SayGoodbye callback.
We only passed the index, previously. Now that we can have multiple
particle systems we need to pass a reference to the particle system,
too.

I already added b2ParticleSystem to the other callbacks, but I missed
the destruction listener.

Tested: Unittests, Testbed, and EyeCandy on Android and Linux.
Bug: 13360060 Particle callbacks need to pass a particle system
pointer.

Change-Id: I74f1c1d5695faf8a4cd91960548c71f24fe9f0a1