Added Multiple Particle Systems test to Testbed.
This test consists of a box with two particle emitters aimed at it, trying
to push the box along a plane.  Each particle emitter is part of a different
particle system so particles from the emitters don't interact with each other.

In addition, fixed a minor bug when creating a particle system with
destroyByAge enabled.

Tested:
Verified the test works on OSX, Linux and Windows.

Bug: 13332986
Change-Id: Ia539d6c29c56e9da20e598543c2fd61ff697cf63