Zipping unittest apks
This change is made to zip the unittests and have them displayed
in the home directory in our build system.
Tested in Linux with creating an archive with -z option.

Removing trailing spaces.
Adding option to specify the tests archive.
Tested on OSX and Linux with -z and -t options.

Change-Id: I8a75d7633e8da9f176195ae4c3bdb958033452ca