Improved preservation of barrier rigid particles.
Since the introduction of SolveRigidDamping(), the momentum of barrier rigid
particles was not preserved. This patch will correct that by exactly evaluating
the velocity of barrier rigid particles.

When a particle is colliding with barrier particles, SolveBarrier() performs the
following two processes.
1) Synchronize velocity of the colliding particle to that of the barrier
particles.
2) Restore the velocity after all particles move.
However, if a rigid particle is colliding, SolveRigid() will erase the effect of
1). That's why momentum won't be conserved.

Tested: ConservationTests on OSX
Bug: 14276652 ConservationTests failling on Linux
Change-Id: I92e56e12f1f3d4d4fc3eea20046486f05dbcce97