Fix Android unit tests.
Set LIQUIDFUN_EXTERNAL_LANGUAGE_API=1 instead of LIQUIDFUN_UNIT_TESTS in
Application.mk. This mirrors the non-Android build process as well in
CMakeLists.txt -- we will always compile in the functions needed for
LIQUIDFUN_EXTERNAL_LANGUAGE_API, so the linker has a choice later.

Change-Id: I145e17157f226c641522d628bb4e0ebb2f59ff50
Tested: Unittests on Linux and Android.