Reorganize b2ParticleContact for easier SIMD access.
Also, compact it a little (indices are now 16-bits instead of 32),
since the contact array is actually quite huge. Reducing the size
improves cache performance non-trivially.

Add equality operators so we can test results between SIMD and
non-SIMD versions, going forward.

Move 'flags' to the bottom of the structure because SIMD instructions
are used on the other fields, and not on the flags field. It's
nice to keep the SIMD data together.

Tested: Testbed on android
Bug: 12192407

Change-Id: I0ca412c932c3d84efca829fe77b06459848028f1