Refactor b2ParticleSystem buffers to use a container class.
This is a small refactoring of the buffer code in b2ParticleSystem.

We put the vector-like buffers into a std::vector-like container.

This is necessary for the SIMD optimization. We need to have a
self-contained self-growable container for b2ParticleContacts.

For now, we only convert those buffers that have variable length.
Most buffers have length equal to the number of particles, and
there's no reason to convert those, at the moment. This change is
big enough as it is!

Change-Id: Ic6a6be1b8df73e80bd08e24233fa07271374d36d
Tested: Unittests on Android and Linux
Bug: 12192407