Extract ContactFilter code from contact test code.
We pass over the contacts an extra time, so this can slow down the
code slightly in the worst case. However, we pass over the contact
array so many times during update, it will always be in L1 cache,
and the slowdown is miniscule.

In the most common case, we have no contact filter, so the code
actually runs slightly faster (since there is no branch in
AddContact).

This change makes it easier to SIMD the contact test, since we're
not doing two things at once.

Tested: Unittests on Linux, Testbed on Android
Bug: 12192407

Change-Id: Ide81460f049144c5660d60e13c86de8ec1fc5847