Refactor contact test code into FindContacts, a const-function.
The const-function is important because we want to run both SIMD
and reference versions and compare the results (for testing).

Also, const-functions are just easier to read, since you know
exactly what they're changing (only the passed in parameter).

The next changelists will start adding SIMD-versions of some of
the refactored functions, including this one.

Change-Id: I45195c1c1d4f8b8d49ebadc31399c998e4ccae6f
Tested: Unittests on Linux and Android, and testbed on Android
Bug: 12192407