Add NEON SIMD version of UpdateProxies.
This version calculates the proxies four at a time, from the
position data, in order of the position data.

This allows us to avoid accessing the position data in proxy-order.
Proxy-order is essentially random and therefore has bad cache
performance. By accessing position data in-order, we can stream
it in using pre-fetching.

After the tags have been calculated, they're used to update the
proxy array.

Add a test of the test AreProxyBuffersTheSame. Yes, a test of the
test!

Before-and-after profiles of the Sandbox test in the Testbed, on
various Android devices. Note that the NEON code currently only
runs on Android, so for all other platforms this change is a nop.

Numbers are usecs per call, on approximately 600 particles with
approximately 1000 collisions per frame.

                         Nexus5 Nexus7 Nexus7 Nexus10 Galaxy
			        (2013) (2012)         Nexus
UpdateProxies_Reference   50     44      45    18      55
UpdateProxies_Simd        13     14      24    12      33

  breakdown of UpdateProxies_Simd:
  CalculateTags_Simd       6      7      10     6      15
  UpdateProxyTags          7      7      14     6      18

Tested: Testbed and Unittests on Android and Linux.
Bug: 12192407
Change-Id: I97df468e39b9fabc32471b031b806a8116c4afa5