Fix compiler warnings when both DEBUG and NDEBUG are defined.
This happens when NDK_DEBUG=0. DEBUG gets defined by us, and NDEBUG
gets defined somewhere in the toolchain.

This combination is a problem because b2Assert ==> assert, but
assert ==> no-op.

The solution is to only enable b2Assert when DEBUG=1 and NDEBUG=0.

There are also a few warnings to fix because DEBUG and
B2_ASSERT_ENABLED are no longer synonymous.

Change-Id: I9303ffa07a8585767fd9d33b6fe97d5c41b4446a
Tested: Testbed on Android, Unittests on Linux