Enforce a minimal alignment on b2StackAllocator allocations.
Problem:
- Allocate 9 bytes
- Next allocation will be 1-byte aligned
- Storing an int32 requires 4-byte alignment ==> crash

The internal malloc enforces 4-byte alignment, so I think it's
reasonable for us to propogate this to b2StackAllocator as well.

Note that this bug has existed for a long time. We simply were
lucky that all our allocation sizes were multiples of 4.

Change-Id: I0d80e2c2c34e4f1f9017b80545e4e003f51f1d53
Tested: Unittests on Android, Linux, Mac.