Move common Android build varibles to b2_android_common.mk.
In a subsequent change, I'll be adding a define for if we use
16 or 32 bits for the particle index in b2ParticleContact.

Since b2ParticleContact is exposed in the external API, we need
to expose this constant to all the applications. The best way to
do this is to have a common makefile that all the applications
reference.

Tested: Unittests, Testbed on Android with ABIs armeabi-v7a and x86
EyeCandy is currently broken from a previous change, but will test
that too when it's fixed.
Bug: 15118924

Change-Id: I9fd1aafcf2ba833e29eece3296552d2f8fc069b0