Corrected timestep for the first step of EyeCandy.
EyeCandy adjusts the length of timestep as follows:
  double time = System_getTime(system_);
  double framedelta = time - prev_time_;
  prev_time_ = time;
  world_->Step(framedelta, 8, 3);

In general 'time' is a large number and currently 'prev_time_' is initialized
as -0.001f, which results in a huge timestep in the first step.

This patch fixes that, so 'prev_time_' will be initialized as
System_getTime(system_).

Bug: 15454100
Change-Id: Ib77d23307cea2e4998ecdef17caa6692ac267127
Tested: on Nexus 5 and iPad 4.