Fix crash after FindContactsFromChecks_Simd.
There was corruption in the array of contacts. This caused all sorts
of trouble, but would mostly manifest as reading off the end of
arrays.

The problem was silly. We save some items to the contact array, but
don't count them in contacts.count until they've been post-processed.
These items weren't being maintained when the contact array was
grown.

The least invasive solution is to always maintain the entire contact
array. This is what we've done to avoid adding instability at this
time.

The better solution is to change FindContactsFromChecks_Simd so that
it stores un-post-processed items in a temporary array. We're
re-using the output array for temporary storage, and it's confusing.

Another problem was that q0~q3 were spilled when
GrowParticleContactBuffer was called. This resulted in garbage being
written out upon return. On the Nexus 10, it appears that memcpy
uses q0, so that's why we were getting crashes on that device.

I misunderstood the calling convention. I assumed only q4~q7 were
spilled, but actually the reverse is true. Only q4~q7 are preserved
across boundaries. I should switch to using q4~q7, but for the
moment, the safer path is just to preserve all the q registers that
I need. That code is called so rarely, it's not a slowdown to push
more data on the stack.

Change-Id: Ia0ed731ba5dd8cd954664776a9b331da12854fdf
Tested: Ran Testbed on Nexus 10, Nexus 7 (2012), Nexus 5.
Bug: 15469518