Fix NaN bug when particles directly on top of each other.
This was a bug in the SIMD function FindContactsFromChecks_Simd.

Detect when 1 / distance is NaN, and set weight to 1 and normal to 0
in that case. This matches the non-SIMD behavior.

Note that detecting when 1 / distance is NaN is more robust than
checking when distance squared is 0. We iterate with reciprocal
square root estimate, so floating point error could turn non-zero
distance squared ==> NaN 1 / distance

Change-Id: I4590d25a7dd31b58b1d8e25f67f459c59f9a343a
Tested: Unittests and Testbed on Android
Bug: 15572206