Change default number of particle iterations to 1.
This is the behavior we had in version 1.0. We're reverting to that
behavior because we were getting, sometimes, large values returned
by b2CalculateParticleIterations.

Added a new helper function to estimate the number of particle
iterations you sould be using:
b2World::CalculationReasonableParticleIterations.

Updated release notes with new API. Update documentation with new
default value.

Change-Id: I370896e0b4f8fbb798cad4b8fa1edc8bafe7fe2b
Tested: Unittests and Testbed on Android and Linux.
Bug: 15538403