Fix Windows bug where mouse input be stuck on one window.
The function GetAsyncKeyState returns two pieces of information.

The first, in the least-significant bit, is whether the key has been
pressed since the last call to GetAsyncKeyState.

The second, in the most-significant bit, is whether the key is
currently pressed.

We want to check only wether the key is pressed. But, we were
checking for either.

The fix is to check only most significant bit.

In the LiquidFun testbed, this bug manifested itself like this:
After the rendering area was pressed, the menu items were no longer
active. The rendering window captured the mouse input and did not
release it until you context-switched away from the testbed.

Change-Id: Ife0a8e72b3b2bf8ed37f57a2d85c58332618025a
Tested: LiquidFun testbed works properly
Bug: 14292657